Tests/LibWeb/Text/expected/Streams/ReadableStream-can-cancel.txt

ReadableStream cancelled due to: Test cancellation
